% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ueda_outliers.R
\name{find_ueda_outliers}
\alias{find_ueda_outliers}
\title{Automatic detection and removal of outliers using Ueda's method.}
\usage{
find_ueda_outliers(x, smax = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{smax}{The maximum number of outliers to be detected. By default, smax = 3.}

\item{na.rm}{Whether NAs should be removed or not for calculating the mean. By default na.rm = TRUE.}
}
\value{
A list of 4 components:

1.  The Ut matrix.

2. The label matrix.

3. The input numeric vector.

4.  The input vector with no outliers.
}
\description{
A function to calculate all components of Ueda's method.
}
\examples{
# random seed for reproducibility
set.seed(13)
# introduce 5 outliers
x <- c(rnorm(25, 300, 10), rnorm(5, 400, 5))
# shuffle the data
x <- sample(x)
# removes up to 5 outliers
find_ueda_outliers(x, smax = 5)
}
\references{
Marmolejo-Ramos, F., Vélez, J.I. & Romão, X. Automatic detection of discordant outliers via the Ueda’s method. J Stat Distrib App 2, 8 (2015). \url{https://doi.org/10.1186/s40488-015-0031-y}
}
