% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ueda_index.R
\name{ueda_index}
\alias{ueda_index}
\title{Find indexes and positions}
\usage{
ueda_index(x, smax = 3)
}
\arguments{
\item{x}{A numeric vector}

\item{smax}{The maximum number of outliers to be detected. By default, smax = 3.}
}
\value{
A list with 2 components:

index  Boolean (TRUE/FALSE) indicating whether each observation is an outlier.

position The position of the outliers in the data.
}
\description{
This function allows identifying the indexes and positions of outliers in a numeric vector.
}
\examples{
# random seed for reproducibility
set.seed(13)
# introduce 5 outliers
x <- c(rnorm(25, 300, 10), rnorm(5, 400, 5))
# shuffle the data
x <- sample(x)
# find indexes
ueda_index(x, smax = 5)
}
\references{
Marmolejo-Ramos, F., Vélez, J.I. & Romão, X. Automatic detection of discordant outliers via the Ueda’s method. J Stat Distrib App 2, 8 (2015). \url{https://doi.org/10.1186/s40488-015-0031-y}
}
